requirements.txt
beautifulsoup4
lxml
tabulate
python-docx
pillow